Vendor import of clang release_39 branch r276489: https://llvm.org/svn/llvm-project/cfe/branches/release_39@276489
Former-commit-id: 9396271fea20393a1c50e7f1bfe2af8b30e334ef

+++ test/CodeGen/const-init.c
@@ -558,7 +558,7 @@ decl_stmt|;
 end_decl_stmt
 
 begin_comment
-comment|// CHECK: @g14 = global i8* inttoptr (i64 100 to i8*)
+comment|// CHECK: @g14 = global i8* inttoptr (i32 100 to i8*)
 end_comment
 
 begin_decl_stmt
